+++ plugins/woocommerce-admin/client/analytics/settings/historical-data/status.js
@@ -5,7 +5,6 @@ import { __, sprintf } from '@wordpress/i18n';
 import { applyFilters } from '@wordpress/hooks';
 import moment from 'moment';
 import { Spinner } from '@wordpress/components';
-import { useFilters } from '@woocommerce/components';
 
 const HISTORICAL_DATA_STATUS_FILTER = 'woocommerce_admin_import_status';
 
@@ -51,6 +50,4 @@ function HistoricalDataStatus( { importDate, status } ) {
 	);
 }
 
-export default useFilters( HISTORICAL_DATA_STATUS_FILTER )(
-	HistoricalDataStatus
-);
+export default HistoricalDataStatus;