+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,6 +1,7 @@
-# Unreleased Changes
+# 5.1.0 (unreleased)
 
-- Fixed default value for `<Table />` component `onQueryChange` prop.
+-   Fixed default value for `<Table />` component `onQueryChange` prop.Fixed default value for `<Table />` component `onQueryChange` prop.
+-   Deprecate our bespoke component `useFilters` in favor of using the WordPress variety `withFilters`.
 
 # 5.0.0